# apiassignment1-sound/apiassignment1-sound/openapi/record-label-openapi-3.1.1.yaml
openapi: 3.1.1
info:
  title: Record Label API
  version: 1.0.0
  description: |
    Record Label API for managing artists.

servers:
  - url: http://localhost:3000

security:
  - basicAuth: []

tags:
  - name: Artists
    description: Artist management

paths:
  /artists:
    get:
      tags: [Artists]
      summary: List artists
      description: Returns a paginated list of artists.
      security:
        - basicAuth: []
      parameters:
        - name: offset
          in: query
          required: false
          description: Page offset (0-based).
          schema:
            type: integer
            minimum: 0
            default: 0
          examples:
            firstPage:
              value: 0
        - name: limit
          in: query
          required: false
          description: Number of items to return.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          examples:
            smallPage:
              value: 2
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistListResponse"
              examples:
                example:
                  value:
                    items:
                      - name: Radiohead
                        genre: Alternative
                        albumsPublished: 9
                        username: radiohead
                    offset: 0
                    limit: 10
                    total: 1
        "400":
          description: Bad Request (invalid pagination parameters)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                invalidLimit:
                  value:
                    detail: "limit must be between 1 and 100"
        "401":
          description: Unauthorized (missing/invalid Basic Auth)
          headers:
            WWW-Authenticate:
              schema:
                type: string
              example: Basic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                unauthorized:
                  value:
                    detail: Unauthorized
    post:
      tags: [Artists]
      summary: Create an artist
      description: Adds a new artist to the database.
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtistCreate"
            examples:
              newArtist:
                value:
                  name: Adele
                  genre: Pop
                  albumsPublished: 4
                  username: adele
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "400":
          description: Bad Request (validation error or duplicate)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          headers:
            WWW-Authenticate:
              schema:
                type: string
              example: Basic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /artists/{artistname}:
    get:
      tags: [Artists]
      summary: Get artist by name
      description: Returns a single artist by their name.
      security:
        - basicAuth: []
      parameters:
        - name: artistname
          in: path
          required: true
          description: The artist name to search for.
          schema:
            type: string
            minLength: 1
          examples:
            example:
              value: Adele
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "401":
          description: Unauthorized
          headers:
            WWW-Authenticate:
              schema:
                type: string
              example: Basic
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                notFound:
                  value:
                    detail: Artist not found

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

  schemas:
    Artist:
      type: object
      additionalProperties: false
      required: [name, genre, albumsPublished, username]
      properties:
        name:
          type: string
          minLength: 1
        genre:
          type: string
          minLength: 1
        albumsPublished:
          type: integer
          minimum: 0
        username:
          type: string
          minLength: 1

    ArtistCreate:
      allOf:
        - $ref: "#/components/schemas/Artist"

    ArtistListResponse:
      type: object
      additionalProperties: false
      required: [items, offset, limit, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Artist"
        offset:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
          minimum: 0

    Error:
      type: object
      additionalProperties: true
      properties:
        detail:
          type: string